requirements.txt
alabaster==0.7.12
appdirs==1.4.4
astroid==2.11.7
Babel==2.10.3
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
colorama==0.4.5
commonmark==0.9.1
coverage==6.4.3
cryptography==37.0.4
dill==0.3.5.1
distlib==0.3.5
docutils==0.16
dotmap==1.3.30
filelock==3.7.1
fire==0.4.0
future==0.18.2
idna==3.3
imagesize==1.4.1
importlib-metadata==4.12.0
isort==5.10.1
Jinja2==3.1.2
keyring==23.8.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
more-itertools==8.13.0
packaging==21.3
pkginfo==1.8.3
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.30
py==1.11.0
pycparser==2.21
Pygments==2.12.0
pylint==2.14.5
pyOpenSSL==22.0.0
pyparsing==3.0.9
pypi-template==0.2.0
python-dotenv==0.21.0
pytz==2022.1
PyYAML==6.0
readme-renderer==36.0
recommonmark==0.7.1
regex==2022.7.25
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.5.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==5.1.1
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-websupport==1.2.4
termcolor==1.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.2
tox==3.25.1
tqdm==4.64.0
twine==4.0.1
typed-ast==1.5.4
typing==3.7.4.3
typing_extensions==4.3.0
urllib3==1.26.11
virtualenv==20.16.3
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
zipp==3.8.1
